requirements.txt
pandas>=2.3.2
numpy>=1.24.0
yfinance>=0.2.31
pandas-ta>=0.3.14b1
pyarrow>=14.0.0
matplotlib>=3.8.0
plotly>=5.18.0
jinja2>=3.1.2
pyyaml>=6.0.1
loguru>=0.7.0
schedule>=1.2.0
requests>=2.31.0
scipy>=1.11.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
anthropic>=0.18.0
openai>=1.12.0
httpx>=0.26.0
aiofiles>=23.2.1
